move[FE]: Move MessageType to separated file
