+++ frontend/src/types/messageType.ts
@@ -1,0 +1,8 @@
+export interface MessageType {
+  id: number;
+  senderId?: number;
+  senderNickname?: string;
+  isLog: boolean;
+  text: string;
+  createdAt: string;
+}
